translate_ai/requirements.txt
transformers
datasets
torch
sentencepiece
accelerate>=0.26.0
fastapi
uvicorn
schedule
protobuf
python-dotenv